requirements.txt
pydantic>=2.0.0
openai>=1.0.0
anthropic>=0.49.0
google-genai>=1.0.0
python-dotenv>=1.0.0
PyYAML>=6.0